requirements.txt
whisper
faiss-cpu
sentence-transformers
ollama
numpy